+++ pkg/providers/property/testdata/TestDSCPCode/match_by_name_and_id.tf
@@ -4,6 +4,7 @@ provider "akamai" {
 
 data "akamai_cp_code" "test" {
   cp_code_name = "test cpcode"
+  cp_code_id   = "cpc_234"
   contract_id  = "ctr_11"
   group_id     = "grp_22"
 }