+++ pkg/providers/property/testdata/TestDSCPCode/match_by_name_output_products.tf
@@ -3,9 +3,9 @@ provider "akamai" {
 }
 
 data "akamai_cp_code" "test" {
-  name        = "test cpcode"
-  contract_id = "ctr_11"
-  group_id    = "grp_22"
+  cp_code_name = "test cpcode"
+  contract_id  = "ctr_11"
+  group_id     = "grp_22"
 }
 
 output "products" {